Moved the last tests
